+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
 ## 2.0.0beta - TBD
+
+**🌟 New**
+- Unordered and ordered nested lists are supported.
+- `allowNestedList` option is added.
+
+**⚠ Breaking changes**
 - `allowFootnote` becomes `allowReference`
 - `onFootnote` becomes `onReference`
+
 
 ## 1.0.0 - 2020-08-13
 - Initial version